restructured command, pattern, and action to be more modular
