+++ rasa/dialogue_understanding/patterns/default_flows_for_patterns.yml
@@ -32,7 +32,7 @@ responses:
         rephrase: True
         template: jinja
 
-  utter_inform_stack_cleanup_rasa:
+  utter_inform_code_change:
     - text: There has been an update to my code. I need to wrap up our running dialogue tasks and start from scratch.
       metadata:
         rephrase: True
@@ -130,11 +130,11 @@ flows:
         next: "start"
       - id: "done"
 
-  pattern_clean_stack:
+  pattern_code_change:
     description: flow used to clean the stack after a bot update
     steps:
       - id: "inform_user"
-        action: utter_inform_stack_cleanup
+        action: utter_inform_code_change
         next: "run_cleanup"
       - id: "run_cleanup"
         action: action_clean_stack
